actions store reducers in utils folder - code refactoring
