+++ client/src/index.js
@@ -3,14 +3,11 @@ import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
-import  store  from "./store";
-import { Provider } from "react-redux";
+
 
 ReactDOM.render(
 	<React.StrictMode>
-		<Provider store={store}>
-			<App />
-		</Provider>
+		<App />
 	</React.StrictMode>,
 	document.getElementById("root")
 );
