+++ client/src/utils/actions.js
@@ -1,0 +1,2 @@
+export const UPDATE_CASES = "UPDATE_CASES";
+
