+++ client/src/utils/store.js
@@ -1,0 +1,3 @@
+import { createStore } from "redux";
+import reducer from "./reducers";
+export default createStore(reducer);
